torrents with files more than int.MaxValue bytes in length are now handled correctly. There shouldn't be any more integer overflowing
svn path=/trunk/bitsharp/; revision=71432

+++ src/MonoTorrent.Client/Managers/PieceManager.cs
@@ -128,7 +128,7 @@ namespace MonoTorrent.Client
         }
 
 
-        internal PieceEvent ReceivedPieceMessage(PeerConnectionID id, byte[] buffer, int dataOffset, int writeIndex, int blockLength, PieceMessage message)
+        internal PieceEvent ReceivedPieceMessage(PeerConnectionID id, byte[] buffer, int dataOffset, long writeIndex, int blockLength, PieceMessage message)
         {
             lock (this.piecePicker)
                 return this.piecePicker.ReceivedPieceMessage(id, buffer, dataOffset, writeIndex, blockLength, message);
